`can_cast` for jax numpy in jax frontend (#7739)
