+++ ivy/functional/frontends/jax/numpy/name_space_functions.py
@@ -6,6 +6,7 @@ from ivy.functional.frontends.jax.func_wrapper import (
     to_ivy_arrays_and_back,
     outputs_to_frontend_arrays,
 )
+from ivy.functional.frontends.jax.numpy import promote_types_of_jax_inputs
 
 
 @to_ivy_arrays_and_back
@@ -18,7 +19,7 @@ abs = absolute
 
 @to_ivy_arrays_and_back
 def add(x1, x2):
-    x1, x2 = ivy.frontends.jax.promote_types_of_jax_inputs(x1, x2)
+    x1, x2 = promote_types_of_jax_inputs(x1, x2)
     return ivy.add(x1, x2)
 
 
@@ -73,9 +74,9 @@ def argsort(a, axis=-1, kind="stable", order=None):
 
 @to_ivy_arrays_and_back
 def asarray(
-        a,
-        dtype=None,
-        order=None,
+    a,
+    dtype=None,
+    order=None,
 ):
     return ivy.asarray(a, dtype=dtype)
 
@@ -102,10 +103,10 @@ def clip(a, a_min=None, a_max=None, out=None):
     )
     a = ivy.array(a)
     if a_min is None:
-        a, a_max = ivy.frontends.jax.promote_types_of_jax_inputs(a, a_max)
+        a, a_max = promote_types_of_jax_inputs(a, a_max)
         return ivy.minimum(a, a_max, out=out)
     if a_max is None:
-        a, a_min = ivy.frontends.jax.promote_types_of_jax_inputs(a, a_min)
+        a, a_min = promote_types_of_jax_inputs(a, a_min)
         return ivy.maximum(a, a_min, out=out)
     return ivy.clip(a, a_min, a_max, out=out)
 
@@ -130,19 +131,18 @@ def cosh(x):
 
 @to_ivy_arrays_and_back
 def dot(a, b, *, precision=None):
-    a, b = ivy.frontends.jax.promote_types_of_jax_inputs(a, b)
+    a, b = promote_types_of_jax_inputs(a, b)
     return ivy.matmul(a, b)
 
 
 @to_ivy_arrays_and_back
 def einsum(
-        subscripts,
-        *operands,
-        out=None,
-        optimize="optimal",
-        precision=None,
-        _use_xeinsum=False
-
+    subscripts,
+    *operands,
+    out=None,
+    optimize="optimal",
+    precision=None,
+    _use_xeinsum=False,
 ):
     return ivy.einsum(subscripts, *operands, out=out)
 
@@ -313,7 +313,7 @@ def flipud(m):
 
 @to_ivy_arrays_and_back
 def power(x1, x2):
-    x1, x2 = ivy.frontends.jax.promote_types_of_jax_inputs(x1, x2)
+    x1, x2 = promote_types_of_jax_inputs(x1, x2)
     return ivy.pow(x1, x2)
 
 
@@ -408,14 +408,14 @@ def kron(a, b):
 
 @to_ivy_arrays_and_back
 def sum(
-        a,
-        axis=None,
-        dtype=None,
-        out=None,
-        keepdims=False,
-        initial=None,
-        where=None,
-        promote_integers=True,
+    a,
+    axis=None,
+    dtype=None,
+    out=None,
+    keepdims=False,
+    initial=None,
+    where=None,
+    promote_integers=True,
 ):
     if initial:
         s = ivy.shape(a)
@@ -616,14 +616,14 @@ def stack(arrays, axis=0, out=None, dtype=None):
 
 @to_ivy_arrays_and_back
 def take(
-        a,
-        indices,
-        axis=None,
-        out=None,
-        mode=None,
-        unique_indices=False,
-        indices_are_sorted=False,
-        fill_value=None,
+    a,
+    indices,
+    axis=None,
+    out=None,
+    mode=None,
+    unique_indices=False,
+    indices_are_sorted=False,
+    fill_value=None,
 ):
     return ivy.take_along_axis(a, indices, axis, out=out)
 
@@ -637,7 +637,7 @@ def zeros_like(a, dtype=None, shape=None):
 
 @to_ivy_arrays_and_back
 def negative(
-        x,
-        /,
+    x,
+    /,
 ):
     return ivy.negative(x)
